Add files via upload
Dice Puzzle similar to those from long ago. Change the dieCount to increase the number of dice and create a more difficult game. The object is to match all of the dice with any number. The die faces will all turn red when you are done. Refresh to restart. Buttons to change levels, restart, etc. coming in a later release.